FuturLearn_main/RAG/requirements.txt
qdrant-client
requests
transformers
torch
python-dotenv
httpx
